Refactor: models into dedicated models file, cosmetics
