+++ cms-ui/apps/admin-ui/src/app/features/mesh-browser/components/mesh-browser-schema-items/mesh-browser-schema-items.component.html
@@ -1,4 +1,3 @@
-
 <div class="schema-items">
 
     <div class="schema-header">
