+++ cms-ui/apps/admin-ui/src/app/features/mesh-browser/components/mesh-browser-schema-list/mesh-browser-schema-list.component.html
@@ -1,4 +1,3 @@
-
 <div class="schema-list-wrapper">
     <ng-container *ngFor="let schema of schemas">
