+++ cms-ui/apps/admin-ui/src/app/features/mesh-browser/components/mesh-browser-schema-list/mesh-browser-schema-list.component.scss
@@ -1,4 +1,3 @@
-
 .schema-list-wrapper{
     margin: 2rem 0;
 }
