+++ cms-ui/apps/admin-ui/src/app/features/mesh-browser/components/mesh-browser-schema-list/mesh-browser-schema-list.component.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Input, OnInit } from '@angular/core';
 import { PaginationService } from 'ngx-pagination';
-import { MeshBrowserLoaderService, Schema, SchemaElement } from '../../providers';
+import { MeshBrowserLoaderService, Schema } from '../../providers';
 
 
 @Component({
