+++ cms-ui/apps/admin-ui/src/app/features/mesh-browser/mesh-browser.module.ts
@@ -17,6 +17,7 @@ import {
     MeshBrowserLoaderService,
 } from './providers';
 
+
 @NgModule({
     declarations: [
         MeshBrowserMasterComponent,
