+++ cms-ui/apps/admin-ui/src/app/features/mesh-browser/providers/mesh-browser-models.ts
@@ -1,0 +1,30 @@
+import { PagingOptions } from '@gentics/mesh-models';
+
+
+export interface MeshSchemaListParams extends PagingOptions  {
+    schemaName?: string,
+    nodeUuid: string,
+}
+
+export interface MeshSchemaListResponse {
+    rootNodeUuid: string,
+    schemas: Array<Schema>
+}
+
+
+export interface Schema {
+    name: string;
+    elements: Array<SchemaElement>;
+}
+
+export interface SchemaElement {
+    displayName: string;
+    uuid: string;
+    isContainer: boolean;
+    languages: Array<ElementLanguage>;
+}
+
+export interface ElementLanguage {
+    language: string;
+}
+
